net/bnx2x: handle guest VLAN for SR-IOV case
In case of bnx2xvf pmd, tx packets can support vland id in 2 ways:
1. setting the mbuf ol_flags=PKT_TX_VLAN_PKT and passing the
vlanid in mbuf->vlan_tci.
2. the tx packet itself has the vlan id included in the packet.
The first case is working as expected but the second case where
the vlan id is included in thetx packets itself was found not
working as expected. To handle that we need to properly set the
start_bd bitfield and the vlan_or_ethertype instead of setting it
to just the ethertype in case of VF.

v3:
 * Fixed the checkpatch issue.
 * Changed vlan to VLAN in the headline.

v2:
 * Fix compilation issues.
 * Changed the Subject Line as recommended.

Signed-off-by: Souvik Dey <sodey@rbbn.com>
Acked-by: Rasesh Mody <rmody@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>