net/i40e: issue with ADD VLAN from Guest
In case of i40evf pmd, when ADD_VLAN is sent down the linux i40e driver,
along with add the vlan the kernel driver also enables the vlan stripping
by default.
This might have issues if the app configured DEV_RX_OFFLOAD_VLAN_STRIP
as off at the port configuration. The app after adding the VLAN will
expect the VLAN to be coming in the received packets but, due to
VLAN_STRIP enabled at the PF, it will get stripped.
This behavior of the Linux driver causes confussion with the DPDK app
using i40e_pmd. So it is better to check the state of the
DEV_RX_OFFLOAD_VLAN_STRIP flag in the dev_conf and disable the vlan strip
after sending add_vlan if app configured the VLAN_STRIP as disabled.

Signed-off-by: Souvik Dey <sodey@rbbn.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>