eal: check for rdseed at run time for random seed
Instead of enabling the rdseed instruction for the random number generator
entropy source at compilation time, determine if the instruction can be
used at run time.

The DPDK build is updated to check that the compiler can generate the
rdseed instruction even if the compilation platform does not natively
support it.  If so, the -mrdseed flag is explicitly added.  This allows
binaries compiled on systems which do not support the rdseed instruction
to use the instruction if the execution platform supports it.

At run-time on x86, __rte_random_initial_seed() will check for the
availability of the rdseed instruction on the execution platform using
rte_cpu_get_flag_enabled().  This allows binaries which were compiled on
systems which support the rdseed instruction to run on x86 CPUs which do
not support the rdseed instruction.

RTE_CPUFLAG_RDSEED is removed from the list of RTE_COMPILE_TIME_CPUFLAGS
which are checked in rte_eal_init() at run time because it is no longer
required to match the compilation system.

Signed-off-by: Dan Gora <dg@adax.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>