eal/trace: define the public API for trace support
Define the public API for trace support.
This patch also adds support for the build infrastructure and
update the MAINTAINERS file for the trace subsystem.

The tracepoint object is of size 8B, it a global variable and used in
fastpath. Created a new __rte_trace_point section to store the
tracepoint objects as,
- It is a mostly read-only data and not to mix with other "write"
global variables.
- Chances that the same subsystem Fast path variables come in the same
fast path cache line. i.e, it will enable a more predictable
performance number from build to build.

Signed-off-by: Jerin Jacob <jerinj@marvell.com>
Signed-off-by: Sunil Kumar Kori <skori@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>