eal: use madvise to exclude unmapped memory from being dumped
Currently, even though memory is mapped with PROT_NONE, this does not
cause it to be excluded from core dumps. This is counter-productive,
because in a lot of cases, this memory will go unused (e.g. when the
memory subsystem preallocates VA space but hasn't yet mapped physical
pages into it).

Use `madvise()` call with MADV_DONTDUMP parameter to exclude the
unmapped memory from being dumped.

Signed-off-by: Li Feng <fengli@smartx.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>