test/service: add perf test for service on app lcore
Add a performance test to the service run on app lcore auto-
test. This test runs the service in a tight loop, and measures
cycles passed, printing the results. It provides a quick cycle
cost value, enabling measuring performance of the function to
run a service on an application lcore.

Signed-off-by: Harry van Haaren <harry.van.haaren@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>