mem: check DMA mask for user-supplied IOVA addresses
Currently, external memory API will silently succeed if the IOVA
addresses supplied by the user do not fit into the DMA mask. This
can cause difficult to debug issues or accepting failed kernel
VFIO DMA mappings.

Fix it so that if the IOVA addresses are provided, they are checked
to see if they fit into the DMA mask.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>