doc: fix references to bind_default_symbol
The document abi_versioning.rst incorrectly instructs the developer to
add BIND_DEFAULT_SYMBOL to the public header, not the source file. This
commit fixes the issue and adds some clarifications.

The commit also clarifies the use of use_function_versioning in the
meson/ninja build system, and does some minor re-organization of the
document.

Signed-off-by: Ray Kinsella <mdr@ashroe.eu>
Signed-off-by: 0-day Robot <robot@bytheb.org>