examples/ipsec-secgw: resolve coverity issue
Function create_ipsec_esp_flow returns a negative number in case of any
failure. But passing negative number to strerror is causing the coverity
issue.
In case of failure, displaying exact error message to console is handled
in create_ipsec_esp_flow function.So it is not required to print the
error message again using strerror.
This patch will remove the unnecessary calling of strerror function
to fix the coverity issue.

Coverity issue: 357691
Fixes: 6738c0a95695 ("examples/ipsec-secgw: support flow director")
Cc: praveen.shetty@intel.com

Signed-off-by: Praveen Shetty <praveen.shetty@intel.com>
Acked-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>