autotest: fix pexpect in python3
In python3 pexpect wants by default a BytesIO stream as log buffer.
Otherwise it will (silently masked since we just return false)
  File "/usr/share/dpdk/test/autotest_runner.py", line 22, in wait_prompt
    child.sendline()
  File "/usr/lib/python3/dist-packages/pexpect/pty_spawn.py", line 555, in sendline
    return self.send(s + self.linesep)
  File "/usr/lib/python3/dist-packages/pexpect/pty_spawn.py", line 543, in send
    self._log(s, 'send')
  File "/usr/lib/python3/dist-packages/pexpect/spawnbase.py", line 126, in _log
    self.logfile.write(s)

We have plenty of other code writing and using this buffer.
So instead of changing the type and modifying many more places we tell the
spawn [1] method to encode things right away - then the rest behaves as it
did with python2.

[1]: https://pexpect.readthedocs.io/en/stable/api/pexpect.html#spawn-class

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>