net/iavf: fix rte flow error log issue
When processing a rte flow, such as creating a parse engine, or
creating or destroying a rss rule, if they are failed, they all
need to construct the flow error structure before return the error
message back to app. If not so, it will cause app crash when
app printing the message out of a flow error.

Fixes: 7be10c3004be ("net/iavf: add RSS configuration for VF")
Fixes: ff2d0c345c3b ("net/iavf: support generic flow API")
Signed-off-by: Jeff Guo <jia.guo@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>