lib/librte_net: fix bug for ipv4 checksum calculating
The function of rte_ipv4_cksum for calculating the
checksum of IPv4 header is incorrect.
This function will return checksum value like 0xffff.
This value, however, is considered an illegal checksum on some switches(like Trident3).

RFC 1624 specifies the IPv4 checksum as follows:
https://tools.ietf.org/rfc/rfc1624
Since there is guaranteed to be at least one
   non-zero field in the IP header, and the checksum field in the
   protocol header is the complement of the sum, the checksum field can
   never contain ~(+0), which is -0 (0xFFFF).  It can, however, contain
   ~(-0), which is +0 (0x0000).

Signed-off-by: 0-day Robot <robot@bytheb.org>