net/mlx5: fix wrong port action assert position
In the current implementation, only one port action is supported in
switchdev mode. So only one port action will exist in a single device
flow. In debug mode, an assertion will be used to check the if the
port action is none, and it should locate before the port action
resource registration.
In the meanwhile, all the fields of the action will be overwritten
in the registration stage. There is no need to clear the temporary
resource with memset.

Fixes: f3faf9ea1185 ("net/mlx5: convert port id action to indexed")
Cc: suanmingm@mellanox.com

Signed-off-by: Bing Zhao <bingz@mellanox.com>
Acked-by: Viacheslav Ovsiienko <viacheslavo@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>