app/testpmd: fix error detection in MTU command
MTU is used in testpmd to set the maximum payload size for packets.
According to testpmd, the setting influnce RX only.
In rte_ethdev there's no relation between MTU setting and JUMBO offload or
rx_max_pkt_len.

The previous fix in patch referenced below was meant to update the
correlated variables of max_pkt_len and JUMBO offload, but by doing
so it assumes that MTU setting can only exist when JUMBO offload
supported in the device. For example fail-safe device does supports set MTU
and doesn't support JUMBO offload, and in this case, though set MTU succeed
an error mesage still printed  since the JUMBO packet offload is disabled.

The fix separates the two conditions to make sure the error
triggers only in case the set_mtu action actually failed.
A warning message is provided in this special case to alert the user.

Fixes: 150c9ac2df13 ("app/testpmd: update Rx offload after setting MTU")
Cc: xavier.huwei@huawei.com

Signed-off-by: Shy Shyman <shys@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>