net/i40e: fix flow FDIR enable issue
When we flush flow FDIR, all queues are disabled for FDIR.
If FDIR rule is created again, then the flow list is empty,
as it is the first time to create rule after flush fdir filter,
so we need to enable FDRI for all queues. And also, disable FDIR
for queues should be done in function i40e_flow_flush_fdir_filter().

Cc: stable@dpdk.org
Fixes: 1491f63c7559 ("net/i40e: fix flush of flow director filter")
Fixes: 6ae9b2b5e8c2 ("net/i40e: cache flow director enable value in Rx queue")

Signed-off-by: Wei Zhao <wei.zhao1@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>