rte_sched: correctly free allocated subport memory
In function rte_sched_subport_free (lib/librte_sched/rte_sched.c,
line 865), there is code to free all allocated stuff related to
scheduler subport. First there are some checks, and in the end,
rte_bitmap_free is called.

Now, rte_bitmap_free is a dummy function, and it just checks if
provided pointer to bitmap is valid or not. So, actual memory for
subport is not freed.

This patch fixes this by removing call to rte_bitmap_free, and
instead calling rte_free.

Signed-off-by: Hrvoje Habjanic <hrvoje.habjanic@zg.ht.hr>
Signed-off-by: 0-day Robot <robot@bytheb.org>