vhost: inroduce operation to get vDPA queue stats
The vDPA device offloads all the datapath of the vhost device to the HW
device.

In order to expose to the user traffic information this patch
introduces new 3 APIs to get traffic statistics, the device statistics
name and to reset the statistics per virtio queue.

The statistics are taken directly from the vDPA driver managing the HW
device and can be different for each vendor driver.

Signed-off-by: Matan Azrad <matan@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>