net/ice/base: add entries in Profile TCAM with priority
The profile TCAM tables are implemented such that entries with a smaller
index in the table have a higher priority. When records to be added to the
table have flags to differentiate between standard PTG and VSIG records,
then these entries need to have higher priority in order to be found and
processed first.

Signed-off-by: Dan Nowlin <dan.nowlin@intel.com>
Signed-off-by: Paul M. Stillwell Jr <paul.m.stillwell.jr@intel.com>
Signed-off-by: Qi Zhang <qi.z.zhang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>